Only strip "Firefox" from UAs for Iceweasel
This is a no-op but helps document why this stripping is here.

The change is required for a follow-up change that integrates the
the "Firefox Mobile" hack into the sequence of wild fixups.
For that to work, we need to be able to enable "Wild fixups" for
Android/Firefox UAs and to not unconditionally strip this word
from the UA string as currently happens.

Fully covered by unit tests already.