test: Separate $.client.test/WikiEditor cases from general UA test
This makes it easier to add new additions by not requiring all
inputs to have dummy matches for what WikiEditor would have done
in 2011, which is hard to justify and basically just means we enter
whatever it happens to produce without it meaning much.

Add a few specific test cases instead to cover the LTR/RTL
functionality, and leave it at that.