test: Swap 'title' and 'userAgent' in test cases
This makes it much more readable by having the input to the test
(userAgent, platform) as siblings.

Previously it was not clear what the purpose of 'title' was
due to being a sibling of 'platform' and 'profile', thus suggesting
it was either an input or an output, when actually it is neither,
it is just the case for the QUnit test case.