[BREAKING CHANGE] Drop Iceweasel distinction, fallback to Firefox
As far as forks go, this is arguably the least significant as it
is literally just a rebranding for policy reasons.

I'm marking this as breaking change, but in practice it most
likely does not require any form of change because:

* It was detected in two places in Code search (VisualEditor
  and MediaWiki core's mediawiki.util), both which only use it
  in order to ensure it is treated the same as Firefox which
  is already detected in both cases as well.

* As of 2016, Iceweasel is end-of-life. The Debian distribution
  that provided it, now uses the Firefox brand fully.

Tests updated accordingly. Note that very old versions
of Iceweasel replaced instead of appended their user agent
string, thus those will now be "unknown" instead of Firefox.

Originally added in 6eec572ae565615, because at the time
Iceweasel did not expose "Firefox" in its UA.

Change-Id: I7f3593fee5528dce7eaea531f2140bb20caf8be0