Simplify wildword replacements, apply always
The wildwords in this pattern effectively matched every modern
browser in existence. The only known browsers not matching the
pattern were legacy MSIE and Konqueror.

This condition was meant as optimisation, but it clearly isn't
right now since it only adds more size and execution cost.

Removing this condition broke MSIE and Konqueror parsing due to
the first wildword blindly removing "MSIE" and "Konqueror" from
all input. From what I can tell, this wasn't needed for anything.
The only issue was that the KHTML/Konqueror needed to include
the slash to avoid false matching.

All tests pass and doing this unconditionally makes the code
easier to reason about I think.

Change-Id: Ibf634ead1f23dcbcdf3757c708b4fbff96da8d26