docs: Add link to npm package
The badge is blocked due to CSP, and we don't yet have a proxy
set up for that.

On Chrome it still renders as a link with alt text, but Firefox
has a non-standard `img:-moz-suppressed { display: none; }` rule
that hides the whole thing... because that's helpful?

Anyway, use a plain link for now.

Change-Id: If227520912baeb619eb076781aca2176375a6144