docs: Improve JSDoc documentation
Introduce a `@typedef` for the output of `$.client.profile()`
(and optional input of `$.client.test()`), which presents almost the
same (a bit more) information in a more organized and machine-readable
way. Use this in the documentation of the private `profileCache` as
well.

Fix the documentation URL to remove the JSDuck-specific hash routing
data, which makes no sense in JSDoc.

Change-Id: I5278692c9121827e6a8ea3578b6b283d01893966